refactor: change utils dir to file to fix missing import
